# page/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Welcome to the Easy Tech API, a cutting-edge startup API provider specializing in a diverse range of services. Explore the world of technology with our APIs, designed to deliver innovative solutions and enhance your applications. Whether you're interested in random anime facts, AI-powered image generation, grammar checks, or historical figure information, Easy Tech API has you covered.\n\nLearn more about Swagger at [http://swagger.io](http://swagger.io) or connect with us on [irc.freenode.net, #swagger](http://swagger.io/irc/).\n",
    "version": "1.0.0",
    "title": "Easy Tech API",
    "contact": {
      "email": "easyapi0@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://api.easy0.repl.co",
      "description": "Easy API"
    }
  ],
...
  "paths": {
...
    "/v1/pixel": {
      "get": {
        "tags": ["SEARCH"],
        "summary": "Search Image",
        "description": "Search Image on Pixelbay.com",
        "operationId": "getLogo",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Provide Your Search Query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "api",
            "in": "query",
            "description": "Api value must be 'ISOYXD'",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "Array"
                }
              }
            }
          }
        }
      }
    },
    "/v2/pixel": {
      "get": {
        "tags": ["SEARCH"],
        "summary": "Search Image",
        "description": "Search Image on Pixels.com",
        "operationId": "v2",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Provide Your Search Query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "Array"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/wallpaper": {
      "get": {
        "tags": ["WALLPAPER"],
        "summary": "Get Wallpaper",
        "description": "Search for wallpapers based on the specified query",
        "operationId": "getWallpaper",
        "parameters": [
          {
            "name": "s",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "url",
                    "example": "https://example.com/wallpaper1.jpg"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Wallpapers not found for the specified query"
          }
        }
      }
    },
    "/v2/wallpaper": {
      "get": {
        "tags": ["WALLPAPER"],
        "summary": "Get Wallpaper",
        "description": "Search for wallpapers based on the specified query",
        "operationId": "wallpaper",
        "parameters": [
          {
            "name": "s",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "url",
                    "example": "https://example.com/wallpaper1.jpg"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Wallpapers not found for the specified query"
          }
        }
      }
    },
...
  }
}